Fix up execution and ensure stdout streaming from subcommands
